# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tao Among API",
        "description": "Tao Among API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/prefix/api/fish": {
            "get": {
                "tags": [
                    "Fish"
                ],
                "summary": "取得魚類列表",
                "operationId": "b13e244d4a771b8009d46e33e736683b",
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            },
            "post": {
                "tags": [
                    "Fish"
                ],
                "summary": "新增魚類",
                "operationId": "9f5210dbcf0bf1a8abc77322f544d10b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "建立成功"
                    }
                }
            }
        },
        "/prefix/api/fish/{id}": {
            "get": {
                "tags": [
                    "Fish"
                ],
                "summary": "取得單一魚類資料",
                "operationId": "151ff128ee6f94ff1cce32f5405b16d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "404": {
                        "description": "找不到資料"
                    }
                }
            },
            "put": {
                "tags": [
                    "Fish"
                ],
                "summary": "更新魚類資料",
                "operationId": "c6eb628750cf59e859da2ce731a77ab8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    },
                    "404": {
                        "description": "找不到資料"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Fish"
                ],
                "summary": "刪除魚類",
                "operationId": "48836c10037d57f5e64d7c14b0ffc64b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "魚類 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刪除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Fish deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "找不到魚類",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "fish not found"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/prefix/api/fish/{id}/notes": {
            "get": {
                "tags": [
                    "Fish"
                ],
                "summary": "取得指定魚類的筆記（可用 since 篩選）",
                "operationId": "a14992bf01c82151212cd99c8c19d14b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "since",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/prefix/api/fish/{id}/note": {
            "post": {
                "tags": [
                    "FishNote"
                ],
                "summary": "新增魚類筆記",
                "operationId": "3d2297c17648d57bd9e0e31b18edab0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "note",
                                    "locate"
                                ],
                                "properties": {
                                    "note": {
                                        "type": "string"
                                    },
                                    "note_type": {
                                        "type": "string"
                                    },
                                    "locate": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "建立成功"
                    }
                }
            }
        },
        "/prefix/api/fish/note/{id}": {
            "put": {
                "tags": [
                    "FishNote"
                ],
                "summary": "更新魚類筆記",
                "operationId": "4811c9adf39794f625fc62d44f3dc2e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "note": {
                                        "type": "string"
                                    },
                                    "note_type": {
                                        "type": "string"
                                    },
                                    "locate": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    },
                    "404": {
                        "description": "找不到該筆記"
                    }
                }
            }
        },
        "/prefix/api/fish/{id}/note/{note_id}": {
            "delete": {
                "tags": [
                    "FishNote"
                ],
                "summary": "刪除魚類筆記",
                "operationId": "b44e4b08d9036b5a88f1093006ab2e61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "魚類 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "note_id",
                        "in": "path",
                        "description": "筆記 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刪除成功"
                    },
                    "404": {
                        "description": "找不到該筆記"
                    }
                }
            }
        },
        "/prefix/api/upload": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "summary": "上傳魚類圖片",
                "description": "上傳魚類圖片",
                "operationId": "a4bcb2e378bd169697ceb21dae229c40",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "要上傳的圖片",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "上傳成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "image uploaded successfully"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "filename.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "驗證失敗",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "驗證失敗"
                                        },
                                        "errors": {
                                            "properties": {
                                                "image": {
                                                    "description": "可能的錯誤訊息：只能上傳單一圖片檔案。|請選擇要上傳的圖片。|圖片格式僅限 jpeg, png, jpg, gif, svg。|圖片大小不可超過 4403 KB。|圖片檔案不可為空。",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "只能上傳單一圖片檔案。"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/prefix/api/get-signed-upload-url": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "summary": "獲取 Supabase 簽名上傳 URL",
                "description": "獲取 Supabase 簽名上傳 URL",
                "operationId": "1ce6f9f6a282b8c86c41d647a02b8202",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "filename"
                                ],
                                "properties": {
                                    "filename": {
                                        "description": "原始檔名",
                                        "type": "string"
                                    },
                                    "path": {
                                        "description": "儲存路徑",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功獲取簽名上傳 URL",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "example": "https://your-project-id.supabase.co/storage/v1/object/public/images/filename.jpg"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "images/filename.jpg"
                                        },
                                        "filename": {
                                            "type": "string",
                                            "example": "filename.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "驗證失敗",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "驗證失敗"
                                        },
                                        "errors": {
                                            "properties": {
                                                "filename": {
                                                    "description": "可能的錯誤訊息：檔名格式不正確。|請提供有效的檔名。",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "檔名格式不正確。"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Fish",
            "description": "Fish"
        },
        {
            "name": "FishNote",
            "description": "FishNote"
        },
        {
            "name": "Upload",
            "description": "Upload"
        }
    ]
}